+++ src/coro/cloudstorage/util/generator_utils.cc
@@ -3,12 +3,12 @@
 namespace coro::cloudstorage::util {
 
 Generator<std::string> Take(Generator<std::string>::iterator& iterator,
-                            int64_t chunk_size) {
+                            size_t chunk_size) {
   while (chunk_size > 0) {
     if ((*iterator).empty()) {
       co_await ++iterator;
     }
-    auto size = std::min<int64_t>((*iterator).size(), chunk_size);
+    auto size = std::min<size_t>((*iterator).size(), chunk_size);
     co_yield std::string((*iterator).begin(), (*iterator).begin() + size);
     chunk_size -= size;
     (*iterator).erase((*iterator).begin(), (*iterator).begin() + size);
